Commonise maxwell3d guest shader caching code
